Add getFitnessFunction for easy access to fitness function
This commit adds a getFitnessFunction method to SingleFitness Algorithms, so the OptimalSolutionStoppingCondition can access the fitness function.
Multi and Many Objective algorithms are currently not supported by the OptimalSolutionStoppingCondition as this would mean, that all single fitness algorithm would either have to keep a list with a single fitness function or constantly create new Lists. This seemed ugly design-wise. Since MOSA and MIO usually do not use the OptimalSolutionStoppingCondition condition anyways I decided not to support them atm.