Replace text in data table that can hardly be internationalised
The text that DataTables dynamically creates cannot be dynamically replaced,
if you do not want to reload the whole page. The solution I implement here
may look trivial but unless you also want to spend hours on this I suggest
we accept this and consider adding a hint somewhere outside of the data table.