Start incorporating TAP13 output into tests table
This is a non-internationalized solution for now as the errors can
have arbitrary fields which are hardly (if at all) internationalizable.
Also, there is a weird issue with errors that are non AssumptionErrors
or AssertionErrors - you somehow cannot treat them like objects.
This results in no report for these kinds of errors (e.g., TypeErrors).