Dockerfile: switch to user "root" as last step
This is necessary to ensure that bind-mounts work properly with our experiment
infrastructure. The host OS must be able to map the user inside a running
container to an existing user outside the container. (In most cases, this is
has to be "root", but our base image uses an unprivileged user "chrome".)
A user other than root might thus not be able to write to the mounted directory.