CI/CD: use image based on Debian Bullseye
Also, specify the Node.JS version manually. The "lts" tag is dynamic and always
refers to the current LTS version of Node.JS. This can silently break things
when a new LTS version is released and the tag refers to the new version, but
the code expects the old version.