feat(TestRunner): fix confusing warning about mismatching seeds
When the generation seed does not match the execution seed, the
TestRunner outputs a warning. However, this warning is also output if
the two seeds are the string "1" and the number 1:

> The generation seed (1) and the execution seed (1) do not match. This may lead to non-deterministic behaviour!

This warning is confusing. Furthermore, we always convert a string seed
into a number seed: convertSeed(). It always tries to parse the seed as
a number first. This means, using "1" or 1 does not make any difference
at all.