test: fix readModelErrors() helper method
It did not work when there was actually an error or a failure in the
model, because it simply grepped for the first occurence of the strings
"modelFails" and "modelErrors". This only worked when there were no
errors, but the strings occur multiple times when there are errors.
The fix is to grep for the first occurence _after_ the "summary" string.