fix: off-by-one in getTotalStepsExecuted() and getTotalTimeExecuted()
getTotalStepsExecuted() is supposed to return the total number of steps
executed by the VM since the test has been started.

There was a subtle behavioral difference to our previous implementation:
In VMWrapper, the step() method executes callbacks before and after
calling the _step() method of the VM's runtime:

```
async step() {
    // ...
    this.callbacks.callCallbacks({afterStep: false}); // (A)
    // ...
    this.vm.runtime._step();
    // ...
    this.callbacks.callCallbacks({afterStep: true}); // (B)
    // ...
}
```

The callbacks executed after _step() are supposed to "see" the old
number of steps taken. That is, getTotalStepsExecuted() should return
the same number both for location (A) and (B). The number of steps taken
should be updated not already after _step(), but only after step().

The same applies to getTotalTimeExecuted().

The commit fixes this.