fix(test-runner.js): let Scratch initialize itself
When an acceleration factor of Infinity is used, it seems that Scratch
(and the renderer in particular) does not have enough time to initialize
itself properly. For example, touchingColor blocks may report false
negatives (e.g., a sprite is reported to not touch a color even though
in reality it is.)

Adding a small delay before starting the test suite seems to fix this.

It also gives Whisker a chance to notify its listeners of the RUN_START
event (i.e., the TestTable and Tap13Listener.)