perf: delay test execution only for dynamic tests
Dynamic tests are ones that encode a neural network that exercises the
program under test. Such networks can be very large compared to
"traditional" static test suites. As such, they also take longer to
load. For this reason, we introduced a default delay of 1 second after
clicking the "runAllTests" button to ensure the network is fully
initialized. But this is not necessary for static tests.

The changes in this commit try to detect when dynamic tests are used (by
looking at the file extension as heuristic) and only delay the execution
if the check is positive.