feat: reset program via save state or reloading
Adds a command line switch and a checkbox in the UI that control if the
Scratch project is reset to its original state between test executions
by either restoring a save state (CLI flag given/checkbox ticked) or
reloading the project (CLI flag not given/checkbox not ticked). The
default is to reload the project.