ci: change scratch-vm dependency if triggered via downstream pipeline
The scratch-vm CI pipeline triggers a downstream build for this project.
It passes the relevant commit hash via the `SCRATCH_VM_COMMIT` variable.
We replace the dependency definition in the `package.json`s accordingly.

For manual pipeline triggers via the GitLab UI, the variable can also be
set to run the CI pipeline against a specific version of the scratch-vm.