fix(run.js): Wait for the project to be uploaded
Otherwise, the #run-all-tests button is clicked too early. When the
_runTestsWithCoverage function initializes the VM, the project is often
not fully loaded into the VM yet. For this reason, vm.runtime.targets is
still an empty array. The CoverageGenerator uses this array to map each
target to block IDs. This is later used to compute the coverage per
target, but also the total coverage. Because the resulting map is also
empty, this coverage computation fails. Effectively, it tries to add
undefined to a number, which results in NaN. The final result is that
NaN will be printed as coverage.