feat(model-editor.js): Always reset array when deleting effects
The comparison against the empty array [] using === does not work,
because it compares object identity, rather than structural equality. If
an emptiness check was desired, the length attribute of the array should
have been compared against 0. As a result, effects was always set to the
empty array. Instead of fixing the bug, I removed the non-working
if-check, because the "buggy" behavior actually makes more sense than
the intended one IMO.