feat: Conditionally disable test table updates
With a large number of tests (> 100) it takes a considerable amount of time (several seconds) to update the test table (e.g., changing the execution status of a test case.) To avoid this overhad, the commit introduces a new flag called "headless" that allows one to conditionally disable these updates. The flag is called "headless" because it's only intended to be used when Whisker Web is run in headless mode -> In this case, the user doesn't see the test table anyway.